Feature/ro 1365 cancel map download

+++ src/app/core/services/offline-map/offline-map.service.ts
@@ -5,7 +5,7 @@ import moment from 'moment';
 import { File, Metadata, Entry } from '@ionic-native/file/ngx';
 import { LoggingService } from '../../../modules/shared/services/logging/logging.service';
 import { BehaviorSubject, from, Observable, Subscription } from 'rxjs';
-import { map, mergeMap } from 'rxjs/operators';
+import { finalize, map, mergeMap } from 'rxjs/operators';
 import { OnReset } from '../../../modules/shared/interfaces/on-reset.interface';
 import { WebView } from '@ionic-native/ionic-webview/ngx';
 import JSZip from 'jszip';
@@ -33,7 +33,7 @@ export class OfflineMapService implements OnReset {
   downloadAndUnzipProgress$ = this.downloadAndUnzipProgress.asObservable();
 
 
-  private downloadAndUnzipSubscriptions: Subscription[] = [];
+  private downloadSubscription: Subscription;
   private cancel = false;
 
   private hasCreatedRootFolder = false;
@@ -87,8 +87,6 @@ export class OfflineMapService implements OnReset {
   }
 
   public async downloadPackage(packageMetadataCombined: CompoundPackageMetadata): Promise<void> {
-    this.cancel = false;
-
     //TODO: Ask user to prefer saving to external SD card if available?
     // TODO: When multiple packages are in queue, the disk space required must also take queued packages into account
     const availableSpace = await this.checkAvailableDiskSpace(packageMetadataCombined.getSizeInMiB());
@@ -102,7 +100,7 @@ export class OfflineMapService implements OnReset {
     // Add new map package to progress subject
     this.downloadAndUnzipProgress.next([...this.downloadAndUnzipProgress.value,  mapPackage]);
 
-    if(this.downloadAndUnzipSubscriptions.length === 0) {
+    if(this.downloadSubscription == null) {
       //Start downlaod if not anyting currently downloading
       this.startDownloadNextItemInQueue();
     }
@@ -132,18 +130,29 @@ export class OfflineMapService implements OnReset {
      this.downloadAndUnzipPart(offlineMapPackage.name, urls[0], urls, offlineMapPackage, 0, parts);
   }
 
-  private cancelDownloadPackage() {
-    this.cancel = true;
-    for(const sub of this.downloadAndUnzipSubscriptions) {
-      sub.unsubscribe();
+  public cancelDownloadPackage(offlineMapPackage: OfflineMapPackage) {
+
+    if(offlineMapPackage.progress.step === ProgressStep.pending) {
+       this.removePackageFromDownloadAndProgress(offlineMapPackage);
+    } else if(offlineMapPackage.progress.step === ProgressStep.download || offlineMapPackage.progress.step === ProgressStep.extractZip) {
+
+      this.cancel = true;
+      this.downloadSubscription?.unsubscribe();
     }
-    this.downloadAndUnzipSubscriptions = [];
+  }
+
+  private removePackageFromDownloadAndProgress(offlineMapPackage: OfflineMapPackage) {
+    this.downloadAndUnzipProgress.next(this.downloadAndUnzipProgress.value.filter(mapPackage => mapPackage.name !== offlineMapPackage.name));
   }
 
   private downloadAndUnzipPart(name: string, url: string, urls: string[], mapPackage: OfflineMapPackage, partNumber: number, totalParts: number) {
     const subfolder = url.indexOf('steepness-outlet') >= 0 ? 'steepness-outlet' : 'statensKartverk'; // Hack to create subfolders in package name folder. For example package 135_74_8 needs to have folders 135_74_8/statensKartverk and 135_74_8/steepness-outlet
     const folder = `${name}/${subfolder}`;
-    this.downloadAndUnzipSubscriptions.push(this.backgroundDownloadService.download(url).subscribe(async (downloadProgress) => {
+    this.downloadSubscription = this.backgroundDownloadService.download(url).pipe(finalize(() => {
+      if(this.cancel) {
+        this.onCancelled(mapPackage);
+      }
+    })).subscribe(async (downloadProgress) => {
       switch(downloadProgress.state) {
         case 'IN_PROGRESS':
           this.onProgress(mapPackage, {step: ProgressStep.download,
@@ -167,7 +176,7 @@ export class OfflineMapService implements OnReset {
         default:
           break;
       }
-    }, (err) => this.onUnzipError(name, err)));
+    }, (err) => this.onUnzipError(name, err));
   }
 
   public calculateTotalProgress(currentProgress: number, currentPart: number, totalParts: number, partOfStep: 'Downloading' | 'Unzipping') {
@@ -179,6 +188,11 @@ export class OfflineMapService implements OnReset {
   }
 
   private onUnzipStepComplete(name: string, urls: string[], mapPackage: OfflineMapPackage, partNumber: number, totalParts: number) {
+    if(this.cancel) {
+      this.onCancelled(mapPackage);
+      return;
+    }
+
     const nextPart = partNumber + 1;
     if(nextPart < totalParts) {
       this.downloadAndUnzipPart(name, urls[nextPart], urls, mapPackage, nextPart, totalParts);
@@ -251,10 +265,13 @@ export class OfflineMapService implements OnReset {
   }
 
   async removeMapPackageByName(packageNameToRemove: string) {
+    // Delete files
     const root = await this.getRootFileUrl();
     await this.deleteDirectory(root, packageNameToRemove);
-    const packages = this.packages.value.filter(mapPackage => mapPackage.name !== packageNameToRemove);
-    this.packages.next(packages);
+    // Remove package from installed packages subject list
+    this.packages.next(this.packages.value.filter(mapPackage => mapPackage.name !== packageNameToRemove));
+    // Also remove any pending / downloading packages in progress
+    this.downloadAndUnzipProgress.next(this.downloadAndUnzipProgress.value.filter(mapPackage => mapPackage.name !== packageNameToRemove));
   }
 
   private async readCompleteFile(packageName: string): Promise<{ size: number, downloadComplete: number }> {
@@ -421,10 +438,8 @@ export class OfflineMapService implements OnReset {
     const unzipProgress = this.downloadAndUnzipProgress.value.filter(p => p.name !== mapPackage.name);
     this.downloadAndUnzipProgress.next(unzipProgress);
 
-    this.clearCurrentDownloadSubscriptions();
-
     // Start any pending downloads
-    this.startDownloadNextItemInQueue();
+    this.resetCancelAndStartNextItemInQueue();
 
     const secondsSpent = mapPackage.downloadComplete - mapPackage.downloadStart;
     const rate = mapPackage.size / 1024 / secondsSpent;
@@ -440,8 +455,15 @@ export class OfflineMapService implements OnReset {
     this.addMapPackage(mapPackage);
   }
 
-  private clearCurrentDownloadSubscriptions() {
-    this.downloadAndUnzipSubscriptions = [];
+  private onCancelled(mapPackage: OfflineMapPackage) {
+    this.removeMapPackageByName(mapPackage.name);
+    this.resetCancelAndStartNextItemInQueue();
+  }
+
+  private resetCancelAndStartNextItemInQueue() {
+    this.cancel = false;
+    this.downloadSubscription = null;
+    this.startDownloadNextItemInQueue();
   }
 
   //#region Directories and file urls
@@ -507,6 +529,10 @@ export class OfflineMapService implements OnReset {
   //#endregion
 
   private async deleteDirectory(path: string, dirName: string): Promise<void> {
+    if(!isAndroidOrIos(this.platform)) {
+      return;
+    }
+
     await this.file
       .removeRecursively(path, dirName)
       .then(() => {
